Removed verify-email api
