+++ app/tests/test_user.py
@@ -25,3 +25,13 @@ class UserTestCase(BaseTestCase):
         login_data = {"email": "user@example.com", "password": "admin"}
         response = self.client.post("/api/login/", data=login_data)
         self.assertEqual(response.status_code, status.HTTP_200_OK)
+
+    def test_onboarding(self):
+        onboarding_data = {
+            "first_name": "Test",
+            "last_name": "Owner",
+            "email": "test_onwer@gmail.com",
+            "organization": {"name": "Test organization", "address": "USA"},
+        }
+        response = self.client.post("/api/owner/onboard/", data=onboarding_data)
+        self.assertEqual(response.status_code, status.HTTP_201_CREATED)
